refactor(image): 임시 이미지 정리 로직을 별도 훅으로 분리
